Allow skipping cgroup path check for container resource limits
Usually the container ID is present in cgroup paths and the code that reads
resource limits from cgroups checks that to avoid treating e.g. root
cgroup settings as per-container values.

In some cases we want to skip the check and read the limits even if
the cgroup path does not contain the container ID.